test: refactor integration tests
