--- tests/integration/helpers/API-service/json/disaster-types.json
@@ -1,52 +1,0 @@
-[
-  {
-    "disasterType": "floods",
-    "label": "flood",
-    "mainExposureIndicator": "population_affected",
-    "showOnlyTriggeredAreas": true,
-    "enableSetWarningToTrigger": false,
-    "leadTimeUnit": "day",
-    "minLeadTime": "0-day",
-    "maxLeadTime": "7-day"
-  },
-  {
-    "disasterType": "malaria",
-    "label": "malaria",
-    "mainExposureIndicator": "potential_cases",
-    "showOnlyTriggeredAreas": false,
-    "enableSetWarningToTrigger": false,
-    "leadTimeUnit": "month",
-    "minLeadTime": "0-month",
-    "maxLeadTime": "2-month"
-  },
-  {
-    "disasterType": "drought",
-    "label": "drought",
-    "mainExposureIndicator": "population_affected",
-    "showOnlyTriggeredAreas": true,
-    "enableSetWarningToTrigger": true,
-    "leadTimeUnit": "month",
-    "minLeadTime": "0-month",
-    "maxLeadTime": "11-month"
-  },
-  {
-    "disasterType": "typhoon",
-    "label": "typhoon",
-    "mainExposureIndicator": "houses_affected",
-    "showOnlyTriggeredAreas": false,
-    "enableSetWarningToTrigger": false,
-    "leadTimeUnit": "hour",
-    "minLeadTime": "0-hour",
-    "maxLeadTime": "168-hour"
-  },
-  {
-    "disasterType": "flash-floods",
-    "label": "flash flood",
-    "mainExposureIndicator": "population_affected",
-    "showOnlyTriggeredAreas": false,
-    "enableSetWarningToTrigger": false,
-    "leadTimeUnit": "hour",
-    "minLeadTime": "0-hour",
-    "maxLeadTime": "48-hour"
-  }
-]
