--- tests/integration/helpers/API-service/json/users.json
@@ -1,90 +1,0 @@
-[
-  {
-    "email": "dunant@redcross.nl",
-    "firstName": "Henry",
-    "lastName": "Dunant",
-    "userRole": "admin",
-    "password": "password"
-  },
-  {
-    "email": "uganda@redcross.nl",
-    "firstName": "Uganda",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["UGA"]
-  },
-  {
-    "email": "zambia@redcross.nl",
-    "firstName": "Zambia",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["ZMB"]
-  },
-  {
-    "email": "zimbabwe@redcross.nl",
-    "firstName": "Zimbabwe",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["ZWE"]
-  },
-  {
-    "email": "kenya@redcross.nl",
-    "firstName": "Kenya",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["KEN"]
-  },
-  {
-    "email": "ethiopia@redcross.nl",
-    "firstName": "Ethiopia",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["ETH"]
-  },
-  {
-    "email": "philippines@redcross.nl",
-    "firstName": "Philippines",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["PHL"]
-  },
-  {
-    "email": "malawi@redcross.nl",
-    "firstName": "Malawi",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["MWI"]
-  },
-  {
-    "email": "malawi-flash-floods@redcross.nl",
-    "firstName": "Malawi",
-    "middleName": "Flash Floods",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["MWI"],
-    "disasterTypes": ["flash-floods"]
-  },
-  {
-    "email": "southsudan@redcross.nl",
-    "firstName": "South Sudan",
-    "lastName": "Manager",
-    "userRole": "local-admin",
-    "password": "password",
-    "countries": ["SSD"]
-  },
-  {
-    "email": "demo-user@redcross.nl",
-    "firstName": "Demo",
-    "lastName": "User",
-    "userRole": "local-admin",
-    "password": "ibf-portal"
-  }
-]
