+++ tests/integration/tests/admin-areas/admin-areas.assertions.ts
@@ -1,0 +1,203 @@
+import { AdminLevel } from '../../helpers/API-service/enum/admin-level.enum';
+import { DisasterType } from '../../helpers/API-service/enum/disaster-type.enum';
+import { LeadTime } from '../../helpers/API-service/enum/lead-time.enum';
+import { FloodsScenario } from '../../helpers/API-service/enum/mock-scenario.enum';
+
+interface Assertion {
+  countryCodeISO3: string;
+  disasterType: DisasterType;
+  adminLevel: AdminLevel;
+  scenario: FloodsScenario;
+  featureCount: number;
+  placeCodeRegex: RegExp;
+  leadTime?: LeadTime;
+  eventName?: string;
+  placeCodeParent?: string;
+}
+
+export const assertions: Assertion[] = [
+  {
+    countryCodeISO3: 'UGA',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 16,
+    placeCodeRegex: /^(UG|21UGA)/, // REFACTOR: set to /^UG/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'UGA',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Warning,
+    featureCount: 2,
+    placeCodeRegex: /^(UG|21UGA)/, // REFACTOR: set to /^UG/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'UGA',
+    disasterType: DisasterType.Drought,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 4,
+    placeCodeRegex: /^(UG|21UGA)/, // REFACTOR: set to /^UG/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'ZMB',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 14,
+    placeCodeRegex: /^ZM/,
+  },
+  {
+    countryCodeISO3: 'ZMB',
+    disasterType: DisasterType.Drought,
+    adminLevel: AdminLevel.adminLevel1,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 10,
+    placeCodeRegex: /^ZM/,
+  },
+  {
+    countryCodeISO3: 'MWI',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel3,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 4,
+    placeCodeRegex: /^(?:MW|Karonga|Blantyre City|Rumphi)/, // REFACTOR: set to /^MW/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'MWI',
+    disasterType: DisasterType.FlashFloods,
+    adminLevel: AdminLevel.adminLevel3,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 3,
+    placeCodeRegex: /^(?:MW|Karonga|Blantyre City|Rumphi)/, // REFACTOR: set to /^MW/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'MWI',
+    disasterType: DisasterType.FlashFloods,
+    adminLevel: AdminLevel.adminLevel3,
+    scenario: FloodsScenario.Trigger,
+    eventName: 'Karonga',
+    featureCount: 7,
+    placeCodeRegex: /^(?:MW|Karonga|Blantyre City|Rumphi)/, // REFACTOR: set to /^MW/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'SSD',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel3,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 1,
+    placeCodeRegex: /^SS/,
+  },
+  {
+    countryCodeISO3: 'KEN',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 4,
+    placeCodeRegex: /^KE/,
+  },
+  {
+    countryCodeISO3: 'KEN',
+    disasterType: DisasterType.Drought,
+    adminLevel: AdminLevel.adminLevel1,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 40,
+    placeCodeRegex: /^KE/,
+  },
+  {
+    countryCodeISO3: 'ETH',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 2,
+    placeCodeRegex: /^ET/,
+  },
+  {
+    countryCodeISO3: 'ETH',
+    disasterType: DisasterType.Malaria,
+    adminLevel: AdminLevel.adminLevel3,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 27,
+    placeCodeRegex: /^ET/,
+  },
+  {
+    countryCodeISO3: 'ETH',
+    disasterType: DisasterType.Drought,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 190,
+    placeCodeRegex: /^ET/,
+  },
+  {
+    countryCodeISO3: 'PHL',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 3,
+    placeCodeRegex: /^PH/,
+  },
+  {
+    countryCodeISO3: 'PHL',
+    disasterType: DisasterType.Typhoon,
+    adminLevel: AdminLevel.adminLevel3,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 60,
+    placeCodeRegex: /^PH/,
+  },
+  {
+    countryCodeISO3: 'ZWE',
+    disasterType: DisasterType.Drought,
+    adminLevel: AdminLevel.adminLevel1,
+    scenario: FloodsScenario.Trigger,
+    featureCount: 3,
+    placeCodeRegex: /^ZW/,
+  },
+  {
+    countryCodeISO3: 'LSO',
+    disasterType: DisasterType.Drought,
+    adminLevel: AdminLevel.adminLevel1,
+    scenario: FloodsScenario.Warning,
+    featureCount: 10,
+    placeCodeRegex: /^LS/,
+  },
+  {
+    countryCodeISO3: 'UGA',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.NoTrigger,
+    featureCount: 128,
+    placeCodeRegex: /^(UG|21UGA)/, // REFACTOR: set to /^UG/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'UGA',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    leadTime: LeadTime.day4,
+    eventName: 'G5220',
+    featureCount: 6,
+    placeCodeRegex: /^(UG|21UGA)/, // REFACTOR: set to /^UG/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'UGA',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel2,
+    scenario: FloodsScenario.Trigger,
+    leadTime: LeadTime.day0,
+    eventName: 'G5075',
+    featureCount: 8,
+    placeCodeRegex: /^(UG|21UGA)/, // REFACTOR: set to /^UG/ after data is fixed
+  },
+  {
+    countryCodeISO3: 'UGA',
+    disasterType: DisasterType.Floods,
+    adminLevel: AdminLevel.adminLevel3,
+    scenario: FloodsScenario.Trigger,
+    leadTime: LeadTime.day0,
+    eventName: 'G5075',
+    placeCodeParent: '21UGA003004',
+    featureCount: 2,
+    placeCodeRegex: /^(UG|21UGA)/, // REFACTOR: set to /^UG/ after data is fixed
+  },
+];
