+++ tests/integration/tests/admin-areas/aggregates.test.ts
@@ -1,46 +1,33 @@
-import * as request from 'supertest';
-
 import { AdminLevel } from '../../helpers/API-service/enum/admin-level.enum';
 import { DisasterType } from '../../helpers/API-service/enum/disaster-type.enum';
 import { FloodsScenario } from '../../helpers/API-service/enum/mock-scenario.enum';
-import { getAccessToken, getServer, mock } from '../../helpers/utility.helper';
-
-export function getAdminAreaAggregates(
-  countryCodeISO3: string,
-  disasterType: DisasterType,
-  adminLevel: AdminLevel,
-  accessToken: string,
-): Promise<request.Response> {
-  return getServer()
-    .get(
-      `/admin-areas/aggregates/${countryCodeISO3}/${disasterType}/${adminLevel}`,
-    )
-    .set('Authorization', `Bearer ${accessToken}`);
-}
+import { mock } from '../../helpers/utility.helper';
+import { getToken } from '../../helpers/utility.helper';
+import { getAdminAreaAggregates } from './admin-areas.api';
 
 export default function adminAreaAggregatesTests() {
-  const countryCodeISO3 = 'UGA';
-  const disasterType = DisasterType.Floods;
-  const adminLevel = AdminLevel.adminLevel3;
-  const scenario = FloodsScenario.Trigger;
-
   describe('admin areas', () => {
-    let accessToken: string;
+    let token: string;
 
     beforeAll(async () => {
-      accessToken = await getAccessToken();
+      token = await getToken();
     });
+
+    const countryCodeISO3 = 'UGA';
+    const disasterType = DisasterType.Floods;
+    const adminLevel = AdminLevel.adminLevel3;
+    const scenario = FloodsScenario.Trigger;
 
     it('should return list of admin area aggregates on GET', async () => {
       // Arrange
-      await mock(scenario, disasterType, countryCodeISO3, null, accessToken);
+      await mock(scenario, disasterType, countryCodeISO3, null, token);
 
       // Act
       const adminAreaAggregates = await getAdminAreaAggregates(
         countryCodeISO3,
         disasterType,
         adminLevel,
-        accessToken,
+        token,
       );
 
       // Assert
